task: move granular role state to parent
